requirements.txt
openai>=1.0.0
faiss-cpu
pdfplumber
langchain
python-dotenv
numpy